Tex files for TOC should now work while in relative directories.
When the table is generated each .aux file is scanned recursively to
find all the tex files. The path of each tex file needed in the table of
contents is calculated with respect to the path of this aux file.

Previously, the tex file path was generated by removing the full path
and the extension. This removed the relative directory. For example,
if main.tex is in the project root directory and you have a tex file
./tex/a.tex in a subdirectory the result is ./a.tex, which does not
exist.

Now, the tex file path is generated by removing the extension (.aux)
and adding .tex.

I have tested this on my own system in two project structures
(flat and multi-level, multi-directory nested) and I have not
encountered a problem.